Tests/LibWeb/Text/expected/WebAudio/BaseAudioContext-decodeAudioData-detached.txt

byteLength: 0
error: DataCloneError
